# openapi.json
{"openapi":"3.0.0","paths":{"/system/languages":{"get":{"operationId":"LanguagesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LanguageDto"}}}}}}},"post":{"operationId":"LanguagesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLanguageDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LanguageDto"}}}}}}},"/system/languages/{code}":{"get":{"operationId":"LanguagesController_findOne","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LanguageDto"}}}}}}},"/system/languages/{id}":{"put":{"operationId":"LanguagesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLanguageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LanguageDto"}}}}}},"delete":{"operationId":"LanguagesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}}}},"/system/voices":{"get":{"operationId":"VoicesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VoiceDto"}}}}}}},"post":{"operationId":"VoicesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVoiceDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceDto"}}}}}}},"/system/voices/{name}":{"get":{"operationId":"VoicesController_findOne","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceDto"}}}}}}},"/system/voices/{id}":{"patch":{"operationId":"VoicesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVoiceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceDto"}}}}}},"delete":{"operationId":"VoicesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/text":{"post":{"operationId":"GenerateTextController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateTextDto"}}}},"responses":{"201":{"description":""}}}},"/profiles":{"post":{"operationId":"UserProfileController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProfileDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}}}},"patch":{"operationId":"UserProfileController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}}}},"get":{"operationId":"UserProfileController_findOne","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}}}}},"/profiles/{id}":{"delete":{"operationId":"UserProfileController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/learn/plans":{"get":{"operationId":"LearnPlansController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LearnPlanDto"}}}}}}},"post":{"operationId":"LearnPlansController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLearnPlanDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearnPlanDto"}}}}}}},"/learn/plans/{id}":{"get":{"operationId":"LearnPlansController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearnPlanDto"}}}}}},"delete":{"operationId":"LearnPlansController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/practice":{"post":{"operationId":"PracticeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePracticeDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeDto"}}}}}},"get":{"operationId":"PracticeController_findAllByLearnPlan","parameters":[{"name":"learnPlanId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PracticeDto"}}}}}}}},"/practice/{practiceTypeId}":{"get":{"operationId":"PracticeController_findByPracticeType","parameters":[{"name":"learnPlanId","required":true,"in":"path","schema":{"type":"string"}},{"name":"practiceTypeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeDto"}}}}}}},"/practice/{id}":{"patch":{"operationId":"PracticeController_update","parameters":[{"name":"learnPlanId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePracticeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeDto"}}}}}},"delete":{"operationId":"PracticeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/system/practice/type":{"post":{"operationId":"PracticeTypeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePracticeTypeDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeType"}}}}}},"get":{"operationId":"PracticeTypeController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PracticeType"}}}}}}}},"/system/practice/type/{id}":{"get":{"operationId":"PracticeTypeController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeType"}}}}}},"patch":{"operationId":"PracticeTypeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePracticeTypeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeType"}}}}}},"delete":{"operationId":"PracticeTypeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Learn Language AI backend API","description":"The API description","version":"1.0","contact":{}},"tags":[{"name":"learn-language-api","description":""}],"servers":[],"components":{"schemas":{"LanguageDto":{"type":"object","properties":{"id":{"type":"number"},"code":{"type":"string"},"name":{"type":"string"}},"required":["id","code","name"]},"CreateLanguageDto":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"}},"required":["name","code"]},"UpdateLanguageDto":{"type":"object","properties":{}},"VoiceDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"gender":{"type":"string"},"languageCode":{"type":"string"}},"required":["id","name","gender","languageCode"]},"CreateVoiceDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"languageCode":{"type":"string"}},"required":["name","gender","languageCode"]},"UpdateVoiceDto":{"type":"object","properties":{}},"AIOptions":{"type":"object","properties":{"modelId":{"type":"string"},"temperature":{"type":"number"},"topP":{"type":"number"},"maxLength":{"type":"number"}},"required":["modelId","temperature","topP","maxLength"]},"GenerateTextDto":{"type":"object","properties":{"textLanguage":{"type":"string"},"nativeLanguage":{"type":"string"},"experience":{"type":"string"},"textDescription":{"type":"string"},"options":{"$ref":"#/components/schemas/AIOptions"}},"required":["textLanguage","nativeLanguage","experience","textDescription","options"]},"CreateProfileDto":{"type":"object","properties":{"userId":{"type":"string"},"nativeLanguage":{"type":"string"}},"required":["userId","nativeLanguage"]},"ProfileDto":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"nativeLanguage":{"type":"string"}},"required":["id","userId","nativeLanguage"]},"UpdateProfileDto":{"type":"object","properties":{"nativeLanguage":{"type":"string"}},"required":["nativeLanguage"]},"LearnPlanDto":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"targetLanguage":{"type":"string"},"level":{"enum":["BEGINNER","INTERMEDIATE","ADVANCED"],"type":"string"}},"required":["id","userId","targetLanguage","level"]},"CreateLearnPlanDto":{"type":"object","properties":{"userId":{"type":"string"},"targetLanguage":{"type":"string"},"level":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]}},"required":["userId","targetLanguage","level"]},"PracticeContentDto":{"type":"object","properties":{"id":{"type":"number"},"input":{"type":"string"},"output":{"type":"object"},"totalTokens":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"practiceId":{"type":"string"}},"required":["id","input","output","totalTokens","createdAt"]},"CreatePracticeDto":{"type":"object","properties":{"practiceType":{"type":"string"},"learnPlan":{"type":"string"},"contentHistory":{"type":"array","items":{"$ref":"#/components/schemas/PracticeContentDto"}}},"required":["practiceType","learnPlan","contentHistory"]},"PracticeTypeDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"instruction":{"type":"string"},"model":{"type":"string"},"temperature":{"type":"number"},"topP":{"type":"number"},"stream":{"type":"boolean"}},"required":["id","name","instruction","model","temperature","topP","stream"]},"PracticeDto":{"type":"object","properties":{"id":{"type":"string"},"practiceType":{"$ref":"#/components/schemas/PracticeTypeDto"},"learnPlan":{"$ref":"#/components/schemas/LearnPlanDto"},"contentHistory":{"type":"array","items":{"$ref":"#/components/schemas/PracticeContentDto"}}},"required":["id","practiceType","learnPlan"]},"UpdatePracticeContentDto":{"type":"object","properties":{}},"UpdatePracticeDto":{"type":"object","properties":{"contentHistory":{"type":"array","items":{"$ref":"#/components/schemas/UpdatePracticeContentDto"}}}},"CreatePracticeTypeDto":{"type":"object","properties":{"name":{"type":"string"},"instruction":{"type":"string"},"model":{"type":"string"},"temperature":{"type":"number","minimum":0,"maximum":2},"topP":{"type":"number","minimum":0,"maximum":1},"stream":{"type":"boolean"}},"required":["name","instruction","model","temperature","topP"]},"PracticeType":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"instruction":{"type":"string"},"model":{"type":"string"},"temperature":{"type":"number"},"topP":{"type":"number"},"stream":{"type":"boolean"}},"required":["id","name","instruction","model","temperature","topP","stream"]},"UpdatePracticeTypeDto":{"type":"object","properties":{}}}}}